Add files via upload
Precision clock on 7 segment LED display, configured by WiFi:
Assembly of a precision clock, with 8 digit 7 segment LED. This display is configured from a mobile phone via WiFi. The date and time information is synchronized through an NTP server, thus making it a very accurate clock. This display is built with the ESP32 module and 8-digit 7-segment LED, being able to connect a second OLED display of 64x48 pixel (0.66 inch).